Factor out common baseclass for remote clusters
